# job-api-project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portail Emploi RESTful",
        "description": "API pour la recherche et la gestion d'offres d'emploi, avec gestion des rôles (Admin, Employeur, Utilisateur).",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Affiche une liste paginée des offres d'emploi (recherche incluse)",
                "operationId": "getJobOffersList",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Recherche par titre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company",
                        "in": "query",
                        "description": "Recherche par entreprise",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Recherche par localisation",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des offres d'emploi."
                    }
                }
            },
            "post": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Crée une nouvelle offre d'emploi (Rôle : employer ou admin requis)",
                "operationId": "createJobOffer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "company",
                                    "location",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "contract_type": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "salary": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre créée avec succès."
                    },
                    "403": {
                        "description": "Accès refusé."
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{jobOffer}": {
            "get": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Affiche une offre d'emploi spécifique",
                "operationId": "getJobOfferById",
                "parameters": [
                    {
                        "name": "jobOffer",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'offre d'emploi."
                    },
                    "404": {
                        "description": "Offre non trouvée ou inactive."
                    }
                }
            },
            "put": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Met à jour une offre d'emploi (Rôle : doit être l'employeur OU admin)",
                "operationId": "updateJobOffer",
                "parameters": [
                    {
                        "name": "jobOffer",
                        "in": "path",
                        "description": "ID de l'offre à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offre mise à jour avec succès."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Job Offers"
                ],
                "summary": "Supprime une offre d'emploi (Rôle : doit être l'employeur OU admin)",
                "operationId": "deleteJobOffer",
                "parameters": [
                    {
                        "name": "jobOffer",
                        "in": "path",
                        "description": "ID de l'offre à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Offre supprimée avec succès."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Personal Access Token",
                "scheme": "bearer"
            }
        }
    },
...
}